backend/requirements.txt
Flask
Flask-CORS
firebase-admin
google-auth
google-auth-oauthlib
google-cloud-dialogflow
